Start implementing changing the clock values
You can now change the year and month (only up, at the moment), and if you
reach month 13, it will be correctly wrapped around.

Actually, there will be lots of code duplication with the calendar app here.
Perhaps it would be good to combine those helper functions in a library and use
it for both apps. Examples of such functions would be nextYear, nextMonth and
so on.