Add the month length computation
The correct month lengths are now used for the wrap-around. One thing that
still has to be done is the leap year computation; right now every year is
taken to not be a leap year.

The functions that do this are generic, and can be moved into the kernel later.